refactor engineSetup to engine
